+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `Request` caching
 * Add `Blast.mkdirp()` and `Blast.mkdirpSync()` methods
 * Add `Blast.rmrf()` and `Blast.rmrfSync()` methods
+* Add temporary files methods: `Blast.generateTempPath()`, `Blast.openTempFile()`, `Blast.cleanupTempPaths()`, ...
 
 ## 0.7.9 (2021-06-11)